ran the script and updated the output
