+++ README.md
@@ -2,37 +2,48 @@
 
 The repository contains the code associated with the paper:
 
-["Micro Risks and (Robust) Pareto Improving Policies with Low Interest Rates"](https://manuelamador.me/files/rlessthang.pdf)
+["Micro Risks and (Robust) Pareto Improving Policies"](https://manuelamador.me/files/rpi.pdf)
     
-by Mark Aguiar, Manuel Amador and Cristina Arellano (2021).
+by Mark Aguiar, Manuel Amador and Cristina Arellano (2024).
+
+Leonardo Barreto provided excellent research assistance and greatly contributed to this code base. 
 
 
 ## Structure
 
 The subfolder `src` contains the main source code.
 
-The subfolder `scripts` contains contains both a julia script as well as the corresponding jupyter notebook that generates the figures and moments reported in the paper.  
+The subfolder `scripts` contains contains the main julia script.  
 
-The subfolder `output` contains the figures generates by the script. 
+The subfolder `output` contains the figures and tables generated by the script. 
 
 ## Running the code 
 
-The code is written in [Julia](https://julialang.org/downloads/). The code uses multithreading if available, so make sure to start julia with the ability to run multiple threads. See [Julia Multi-threading](https://docs.julialang.org/en/v1/manual/multi-threading/#man-multithreading).
+The code is written in [Julia](https://julialang.org/downloads/), tested on **Version 1.10.4 (2024-06-04)**.
+
+IMPORTANT: The code uses multithreading, if available, for a siginificant speed up, so make sure to start julia with the ability to run multiple threads. To let Julia decide the number of threads to use, start with the option `-t auto`:
+
+    $ julia -t auto
+
+See [Julia Multi-threading](https://docs.julialang.org/en/v1/manual/multi-threading/#man-multithreading) for details.
 
 To run the code, open a julia prompt at the root of this repository. 
 Then type:
 
-    julia> using Pkg 
-    julia> Pkg.activate(".")
-    julia> Pkg.instantiate()
+    julia> include(joinpath("scripts", "main.jl"))
 
-The above will download the Julia packages needed to run the programs.
+The script will 
 
-To run the main script that generates the figures and statistics, do:
-  
-    julia> include(joinpath("scripts", "paper_figures_tables.jl")) 
-    
-The script takes 40 minutes to run a in a server with AMD EPYC 7542 32-Core Processor with 20 cores available and 32GB of RAM.  
+- Automatically download and install all the necessary packages.
 
-There is a Jupyter notebook associated with the script that can be run instead. Run [Jupyter](https://jupyter.org/) and navigate to `scripts` to locate the notebook and run it.
+- Delete all of the output in the "output" directory. 
 
+- Repopulate the output directory with the figures and tables.
+ 
+The script (once the necessary packages are installed) takes approximately 30 minutes to run on a Surface Pro 8 laptop with 
+
+- Processor: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz, 2995 Mhz, 4 Core(s), 8 Logical Processor(s), 
+
+- 16GB of RAM, 
+
+- Running Windows 11 Pro.
