+++ output/tables/statistics.txt
@@ -1,0 +1,25 @@
+Steady State Statistics
+┌───────────────────────┬───────────────┬────────────┬──────────┐
+│                       │ Laissez Faire │ Constant K │ Golden K │
+├───────────────────────┼───────────────┼────────────┼──────────┤
+│               b/y (%) │             0 │         60 │       56 │
+│                 r (%) │          -1.7 │       -1.4 │     -1.1 │
+│                   k/y │           2.6 │        2.6 │      3.0 │
+│               T/y (%) │           0.0 │        0.1 │      0.3 │
+│ Shared constrained hh │             3 │          2 │        2 │
+│       Q1 Wealth Share │             1 │          1 │        1 │
+│       Q2 Wealth Share │             4 │          4 │        5 │
+│       Q3 Wealth Share │            10 │         11 │       11 │
+│       Q4 Wealth Share │            23 │         23 │       24 │
+│       Q5 Wealth Share │            63 │         61 │       60 │
+└───────────────────────┴───────────────┴────────────┴──────────┘
+Welfare Gains at announcement
+┌────────────────┬────────────┬───────────────────┐
+│                │ Constant K │ Capital Expansion │
+├────────────────┼────────────┼───────────────────┤
+│           mean │        2.6 │               5.2 │
+│            min │        2.1 │               4.5 │
+│     poor (≤10) │        3.7 │               5.3 │
+│ middle (40-60) │        2.3 │               4.8 │
+│     rich (>90) │        2.8 │               6.7 │
+└────────────────┴────────────┴───────────────────┘
