+++ output/tables/statistics_all.txt
@@ -1,0 +1,23 @@
+Table 2 Baseline Constant K Policy and Laissez Faire Economies
+┌─────────────────┬────────────┬───────────────┐
+│                 │ Constant K │ Laissez Faire │
+├─────────────────┼────────────┼───────────────┤
+│         b/y (%) │         60 │             0 │
+│           r (%) │       -1.4 │          -1.7 │
+│             k/y │        2.6 │           2.6 │
+│ Q1 Wealth Share │          1 │             1 │
+│ Q2 Wealth Share │          4 │             4 │
+│ Q3 Wealth Share │         11 │            10 │
+│ Q4 Wealth Share │         23 │            23 │
+│ Q5 Wealth Share │         61 │            63 │
+└─────────────────┴────────────┴───────────────┘
+Table 1 Changes in Welfare
+┌────────────────┬────────────┬───────────────────┬─────────────┬───────────────────┐
+│                │ Constant K │ High Initial Debt │ Agg. Shocks │ Capital Expansion │
+├────────────────┼────────────┼───────────────────┼─────────────┼───────────────────┤
+│           mean │        2.6 │               0.7 │         2.7 │               5.2 │
+│            min │        2.1 │               0.5 │         2.1 │               4.5 │
+│     poor (≤10) │        3.7 │               1.0 │         3.6 │               5.3 │
+│ middle (40-60) │        2.3 │               0.6 │         2.4 │               4.8 │
+│     rich (>90) │        2.8 │               0.8 │         2.8 │               6.7 │
+└────────────────┴────────────┴───────────────────┴─────────────┴───────────────────┘
